requirements.txt
black
flake8
pandas
requests
biopython
scikit-learn
torch
torchtext
seaborn
fastparquet
numpy
selenium
loguru
tf-keras
accelerate
datasets
transformers
torch-geometric
mdtraj
Esm
huggingface_hub
fsspec
pykan
dssp
pyyaml
